- ChannelAdvisor: Update to checking if Cost has value - Product/Location: Implemented pulling product locations and quantity   from Skuvault - Sale/Cart: Refactored form to prevent unnecessary properties from   being bound with the model - AddRemoveItemBulkModel: Model for use with adding and removing   products in bulk. Allows for more flexible use with other scenarios   rather than having to create additional methods in SkuVault Library - Sale/Refund: Similar update as Sale/Cart - SaleLine: Updated discount columns to be float/money instead of int - Bugs/Products_Update: Fixed spelling error of Update to Updated - SkuVaultLibrary: Updated how warehouse Ids are retrieved since API   only allows one request per minute. It'll now store the warehouse ID.   (Temporarily hardcoding for testing purpose) - SkuVault/RemoveItemBulkAsync: Refactored to reflect   AddRemoveItemBulkModel and properly distinguish between SKU and UPC - SkuVault/AddItemBulkAsync: Implemented adding products. This   combined with RemoveItemBulkAsync will allow for product transfers - SkuVault/GetInventoryLocationsAsync: Implemented method to retrieve   available locations and quantities

+++ NewAgePOSDB/dbo/StoredProcedures/spSaleLines_Update.sql
@@ -1,8 +1,8 @@
 ﻿CREATE PROCEDURE [dbo].[spSaleLines_Update]
 	@id int,
 	@qty int,
-	@discAmt int,
-	@discPct int
+	@discAmt float,
+	@discPct float
 AS
 BEGIN
 	SET NOCOUNT ON;
